Move `CodableGalleryStore` to production and change access control
