+++ NASAGallery/NASAGalleryTests/Gallery Cache/Infrastructure/CodableGalleryStoreTests.swift
@@ -35,71 +35,6 @@ import NASAGallery
 
 */
 
-final class CodableGalleryStore: GalleryStore {
-    let storeURL: URL
-    
-    init(storeURL: URL) {
-        self.storeURL = storeURL
-    }
-    
-    func retrieve() throws -> LocalCache? {
-        guard FileManager.default.fileExists(atPath: storeURL.path()) else { return nil }
-        
-        let data = try Data(contentsOf: storeURL)
-        let cache = try JSONDecoder().decode(Cache.self, from: data)
-        return LocalCache(gallery: cache.localGallery, timestamp: cache.timestamp)
-    }
-    
-    func insert(_ cache: LocalCache) throws {
-        let codableCache = Cache(gallery: cache.gallery.map(CodableLocalGalleryImage.init), timestamp: cache.timestamp)
-        let data = try JSONEncoder().encode(codableCache)
-        try data.write(to: storeURL)
-    }
-    
-    func deleteCachedGallery() throws {
-        guard FileManager.default.fileExists(atPath: storeURL.path()) else { return }
-        try FileManager.default.removeItem(at: storeURL)
-    }
-    
-    // MARK: - DTOs
-    
-    private struct Cache: Codable {
-        let gallery: [CodableLocalGalleryImage]
-        let timestamp: Date
-        
-        var localGallery: [LocalGalleryImage] {
-            return gallery.map { $0.local }
-        }
-    }
-
-    private struct CodableLocalGalleryImage: Codable {
-        let title: String
-        let url: URL
-        let date: String
-        let explanation: String
-        let mediaType: String
-        
-        let copyright: String?
-        let hdurl: URL?
-        let thumbnailUrl: URL?
-        
-        public init(local: LocalGalleryImage) {
-            self.title = local.title
-            self.url = local.url
-            self.date = local.date
-            self.explanation = local.explanation
-            self.mediaType = local.mediaType
-            self.copyright = local.copyright
-            self.hdurl = local.hdurl
-            self.thumbnailUrl = local.thumbnailUrl
-        }
-        
-        var local: LocalGalleryImage {
-            LocalGalleryImage(title: title, url: url, date: date, explanation: explanation, mediaType: mediaType, copyright: copyright, hdurl: hdurl, thumbnailUrl: thumbnailUrl)
-        }
-    }
-}
-
 final class CodableGalleryStoreTests: XCTestCase {
     
     // MARK: - SetUp & TearDown
